templates/LlamaIndex_RAG_service_lab/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["matamong <matamatamong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.0"}
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
gunicorn = "^21.2.0"
rich = "^13.7.1"
httpx = "^0.27.0"
llama-index-llms-ollama = "^0.1.5"
llama-index-embeddings-huggingface = "^0.2.1"
llama-index-readers-file = "^0.1.23"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
